pyproject.toml
[project]
name = "karton-yara-matcher"
version = "0.1.0"
description = "Karton service to run YARA on samples"
readme = "README.md"
authors = [{ name = "Kevin Ziadeh" }]
requires-python = ">=3.12"
dependencies = [
    "karton-core>=5.6.0",
    "yara-python>=4.5.4",
]
classifiers = [ 
    "Private :: Do not Upload",
    "Operating System :: Unix",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "S603", # subprocess.check_output
    "G004", # f-string log strings
    "TRY003", # long strings outside exception class
    "EM101", # strings in error raise
    "EM102", # f-strings in error raise
    "D100", # missing docstring in public module
    "D104", # missing docstring in public package
    "D212", # multi-line docstring summary should start at the first line
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "N999"]

[project.scripts]
karton-yara-matcher = "src.yara_matcher:YaraMatcher.main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "ruff>=0.9.8",
]
